moved filterUrlLink location
